casted socket.size() to uint64_t because %lu for size_t still generates a warning

+++ clcomm.cpp
@@ -13,6 +13,8 @@
 #include <SDL/SDL.h>
 #include <cstdio>
 #include <cmath>
+#include <stdint.h>
+
 using namespace std;
 
 #define MYPORT "55555"
@@ -108,7 +110,7 @@ int main() {
 					for(int j = i; j < clients.size() - 1; j++)
 						clients[j] = clients[j+1];
 					clients.pop_back();
-					printf("Client disconnected size = %lu\n", clients.size());
+					printf("Client disconnected size = %lu\n", (uint64_t)clients.size());
 					i--;
 					break;
 				}
